Fix non-predictable exit code.
The exit code used by die can be non-predictable. This is a problem
because nagios scripts are expected to return specific exit codes. This
is fixed by using eval around die. This permits to set a specific exit
code.

The option "--no-stats-exit-code" is added so the user can choose which
exit code to use when it is not possible to obtain the haproxy stats. By
default this value is set to 3 (state UNKNOWN).

Changed exit code for help from 3 to 0 because it is a clean exit
without errors and common for nagios scripts.